services/worker/requirements.txt
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.1
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
httpx==0.27.0
openpyxl==3.1.2
cryptography==42.0.4
